+++ ketcher-autotests/tests/utils/selectors/index.ts
@@ -2,4 +2,3 @@ export * from './buttons';
 export * from './templateModal';
 export * from './attachment-point-modal';
 export * from './selectStructures';
-export { selectTemplate } from './selectTemplate';
